longオプションに対応するため既存の処理をモジュールに分割した
