+++ 05.ls/ls2.rb
@@ -2,10 +2,14 @@
 # frozen_string_literal: true
 
 require 'optparse'
+require_relative 'ls_short'
 
 class Ls
-  COLS = 3 # 列数
-  SPAN = 2 # 列間のスペース数
+  include LsShort
+
+  def self.show(target, options)
+    new.show(target, options)
+  end
 
   def show(target, options)
     if target && File.file?(target)
@@ -27,33 +31,6 @@ class Ls
     Dir.chdir(dir)
     Dir.entries('.').sort_by { |v| v.match('^[.]?(.*$)')[1] }
   end
-
-  def generate_table(items)
-    rows = (items.size.to_f / COLS).ceil
-    horizontal_table = items.each_slice(rows).to_a
-    aligned_horizontal_table = horizontal_table.map do |col_data|
-      width = col_data.map { |item| digit_size(item) }.max + SPAN
-      col_data.map { |item| digit_ljust(width, item) }
-    end
-    aligned_horizontal_table.map { |item| item.values_at(0...rows) }.transpose
-  end
-
-  def print_table(table)
-    table.each do |row_data|
-      row_data.compact.each { |item| print item }
-      print "\n"
-    end
-  end
-
-  # 以下は全角文字に対応するためのメソッド
-  def digit_size(text)
-    text.each_char.map { |c| c.bytesize == 1 ? 1 : 2 }.inject(:+)
-  end
-
-  def digit_ljust(width, text)
-    padding = ' ' * (width - digit_size(text))
-    text + padding
-  end
 end
 
 options = {}
@@ -62,5 +39,4 @@ opt.on('-a') { |v| options[:a] = v }
 opt.on('-r') { |v| options[:r] = v }
 target = opt.parse(ARGV)[0] # ファイル/ディレクトリ指定2つ目以降は無視
 
-ls = Ls.new
-ls.show(target || '.', options)
+Ls.show(target || '.', options)
